+++ docs/CHANGELOG.rst
@@ -50,6 +50,7 @@ Added
 - Add ``workflow-accel-gatk4`` pipeline that uses GATK4 HaplotypeCaller as an
   alternative to GATK3 used in ``workflow-accel`` pipeline
 - Add ``amplicon-master-file`` descriptor schema
+- Add ``workflow-bbduk-star-featurecounts`` pipeline
 
 Fixed
 -----